Adding condition functionality
Conditions have been changed significantly. The previous version only offered true/false binary choices, and while that works for many simple objects, this will work for both the simple cases and the more complex ones. This is also the reason why I made it a separate class from the beginning.