Refactored buildItems and added item storage to Game
Making some smaller changes before I push the next PR with the "use" verb included.

I added an item storage list to the Game class to hold items which need to be accessed later, as well as methods for adding/removing. This keeps the creation of items all in one place.

I also refactored the buildItems function per the discussion in PR #71